requirements.txt
# Core dependencies (compatible with Python 3.13)
numpy>=1.19.0
matplotlib>=3.3.0
Pillow>=8.0.0

# Optional dependencies for enhanced functionality
opencv-python>=4.5.0
scikit-image>=0.18.0
tqdm>=4.60.0

# For plotting and visualization
seaborn>=0.11.0

# Note: TensorFlow installation may require Python 3.8-3.11
# For Python 3.13, consider using:
# - TensorFlow CPU: pip install tensorflow-cpu
# - Or downgrade to Python 3.11 for full TensorFlow support